rename real_app_test+commends file
